api/requirements.txt
aniso8601==9.0.1
annotated-types==0.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.4
dnspython==2.3.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Pydantic==0.11.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
pymongo==4.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
typing_extensions==4.7.1
watchdog==3.0.0
Werkzeug==2.3.6
